# openapi/personnel-administration-api.yaml
openapi: 3.0.3
info:
  title: B4RRHH Personnel Administration API
  version: 0.1.0
...
paths:
...
  /employees/{ruleSystemCode}/{employeeTypeCode}/{employeeNumber}/identifiers:
    post:
      operationId: createIdentifierByBusinessKey
      summary: Create employee identifier
      tags:
        - Employee Identifier
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: employeeTypeCode
          required: true
          schema:
            type: string
            maxLength: 30
        - in: path
          name: employeeNumber
          required: true
          schema:
            type: string
            maxLength: 15
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIdentifierRequest"
      responses:
        "201":
          description: Identifier created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentifierResponse"
        "400":
          description: Invalid request
        "404":
          description: Rule system, employee, or catalog not found
        "409":
          description: Identifier already exists for type or primary conflict
    get:
      operationId: listEmployeeIdentifiersByBusinessKey
      summary: List employee identifiers
      tags:
        - Employee Identifier
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: employeeTypeCode
          required: true
          schema:
            type: string
            maxLength: 30
        - in: path
          name: employeeNumber
          required: true
          schema:
            type: string
            maxLength: 15
      responses:
        "200":
          description: Identifier list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IdentifierResponse"
        "404":
          description: Rule system or employee not found
  /employees/{ruleSystemCode}/{employeeTypeCode}/{employeeNumber}/identifiers/{identifierTypeCode}:
    get:
      operationId: getIdentifierByBusinessKey
      summary: Get employee identifier by business key
      tags:
        - Employee Identifier
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: employeeTypeCode
          required: true
          schema:
            type: string
            maxLength: 30
        - in: path
          name: employeeNumber
          required: true
          schema:
            type: string
            maxLength: 15
        - in: path
          name: identifierTypeCode
          required: true
          schema:
            type: string
            maxLength: 30
      responses:
        "200":
          description: Identifier found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentifierResponse"
        "404":
          description: Rule system, employee, or identifier not found
    put:
      operationId: updateIdentifierByBusinessKey
      summary: Update employee identifier
      tags:
        - Employee Identifier
      parameters:
        - $ref: "#/components/parameters/AcceptLanguage"
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: employeeTypeCode
          required: true
          schema:
            type: string
            maxLength: 30
        - in: path
          name: employeeNumber
          required: true
          schema:
            type: string
            maxLength: 15
        - in: path
          name: identifierTypeCode
          required: true
          schema:
            type: string
            maxLength: 30
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIdentifierRequest"
      responses:
        "200":
          description: Identifier updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentifierResponse"
        "400":
          description: Invalid request
        "404":
          description: Rule system, employee, catalog, or identifier not found
        "409":
          description: Primary identifier conflict
    delete:
      operationId: deleteIdentifierByBusinessKey
      summary: Delete employee identifier
      tags:
        - Employee Identifier
      parameters:
        - in: path
          name: ruleSystemCode
          required: true
          schema:
            type: string
            maxLength: 5
        - in: path
          name: employeeTypeCode
          required: true
          schema:
            type: string
            maxLength: 30
        - in: path
          name: employeeNumber
          required: true
          schema:
            type: string
            maxLength: 15
        - in: path
          name: identifierTypeCode
          required: true
          schema:
            type: string
            maxLength: 30
      responses:
        "204":
          description: Identifier deleted
        "404":
          description: Rule system, employee, or identifier not found
...
components:
  parameters:
    AcceptLanguage:
      in: header
      name: Accept-Language
      required: false
      description: >-
        Language of the catalog labels in the response (ADR-052). Short BCP 47 tag such as
        `es-ES`, `fr-FR` or `en`; the heaviest language of the header is used. Codes never
        change. A label without a translation for that language, a missing header or an
        unrecognised one all fall back to the base literal, silently.
      schema:
        type: string
        example: es-ES
  schemas:
...
    CreateIdentifierRequest:
      type: object
      additionalProperties: false
      required:
        - identifierTypeCode
        - identifierValue
      properties:
        identifierTypeCode:
          type: string
          maxLength: 30
        identifierValue:
          type: string
          maxLength: 120
        issuingCountryCode:
          type: string
          maxLength: 3
          nullable: true
          description: Optional ISO 3166-1 alpha-3 code validated against active COUNTRY rule entities for the employee ruleSystemCode.
        expirationDate:
          type: string
          format: date
          nullable: true
        isPrimary:
          type: boolean
          nullable: true
    UpdateIdentifierRequest:
      type: object
      additionalProperties: false
      required:
        - identifierValue
      properties:
        identifierValue:
          type: string
          maxLength: 120
        issuingCountryCode:
          type: string
          maxLength: 3
          nullable: true
          description: Optional ISO 3166-1 alpha-3 code validated against active COUNTRY rule entities for the employee ruleSystemCode.
        expirationDate:
          type: string
          format: date
          nullable: true
        isPrimary:
          type: boolean
          nullable: true
    IdentifierResponse:
      type: object
      required:
        - identifierTypeCode
        - identifierValue
        - isPrimary
      properties:
        identifierTypeCode:
          type: string
        identifierTypeName:
          type: string
          nullable: true
          description: Optional display name from EMPLOYEE_IDENTIFIER_TYPE rule entities in the employee ruleSystemCode catalog.
        identifierValue:
          type: string
        issuingCountryCode:
          type: string
          nullable: true
          description: Optional ISO 3166-1 alpha-3 code from COUNTRY rule entities in the employee ruleSystemCode catalog.
        expirationDate:
          type: string
          format: date
          nullable: true
        isPrimary:
          type: boolean
...